--- src/main/java/com/hearthstone/hearthstone_deck_storage/HearthstoneDeckStorageApplication.java
@@ -1,13 +1,0 @@
-package com.hearthstone.hearthstone_deck_storage;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HearthstoneDeckStorageApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(HearthstoneDeckStorageApplication.class, args);
-    }
-
-}
