--- src/main/java/com/hearthstone/hearthstone_deck_storage/api/token/OAuth2FlowHandler.java
@@ -1,10 +1,0 @@
-package com.hearthstone.hearthstone_deck_storage.api.token;
-
-import com.hearthstone.hearthstone_deck_storage.api.token.model.TokenResponse;
-
-public interface OAuth2FlowHandler {
-
-    TokenResponse getToken();
-
-    boolean isTokenInvalid();
-}
