--- src/main/java/com/hearthstone/hearthstone_deck_storage/api/token/model/TokenResponse.java
@@ -1,12 +1,0 @@
-package com.hearthstone.hearthstone_deck_storage.api.token.model;
-
-import lombok.Data;
-
-import java.time.Instant;
-
-@Data
-public class TokenResponse {
-    private String token_type;
-    private String access_token;
-    private Instant expires_in;
-}
