--- src/main/java/com/hearthstone/hearthstone_deck_storage/common/config/ApiClientConfig.java
@@ -1,6 +1,0 @@
-package com.hearthstone.hearthstone_deck_storage.common.config;
-
-public interface ApiClientConfig {
-    String getBaseUrl();
-    String getTokenUrl();
-}
