--- src/main/java/com/hearthstone/hearthstone_deck_storage/common/config/EnvConfig.java
@@ -1,13 +1,0 @@
-package com.hearthstone.hearthstone_deck_storage.common.config;
-
-import lombok.Data;
-
-@Data
-public class EnvConfig {
-
-    public static final String CLIENT_ID = "ba3276ab8cc24d4fbe6ed3b8e4743bf2";
-    public static final String CLIENT_SECRET = "bkwrkKkx7IxkdlaW3PnqVwQMs264Q97W";
-    public static final String TOKEN_URL_NAME = "https://oauth.battle.net/token";
-    public static final String BASE_URL_NAME = "https://eu.api.blizzard.com/hearthstone/cards";
-
-}
