--- src/main/java/com/hearthstone/hearthstone_deck_storage/common/converter/CharacterClassConverter.java
@@ -1,26 +1,0 @@
-package com.hearthstone.hearthstone_deck_storage.common.converter;
-
-import com.hearthstone.hearthstone_deck_storage.dao.entity.enums.CharacterClass;
-
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
-
-@Converter
-public class CharacterClassConverter implements AttributeConverter<CharacterClass, Integer> {
-    @Override
-    public Integer convertToDatabaseColumn(CharacterClass attribute) {
-        if (attribute == null) {
-            return null;
-        }
-        return attribute.getValue();
-    }
-
-    @Override
-    public CharacterClass convertToEntityAttribute(Integer dbData) {
-        if (dbData == null) {
-            return null;
-        }
-        return CharacterClass.fromValue(dbData);
-    }
-
-}
