--- src/main/java/com/hearthstone/hearthstone_deck_storage/service/CardService.java
@@ -1,21 +1,0 @@
-package com.hearthstone.hearthstone_deck_storage.service;
-
-import com.hearthstone.hearthstone_deck_storage.dto.CardDto;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface CardService {
-
-    CardDto getOrSave(CardDto cardDto);
-
-    List<CardDto> saveAll(List<CardDto> cardDtoList);
-
-    Optional<CardDto> findById(Long id);
-
-    List<CardDto> findAll();
-
-    void delete(CardDto cardDto);
-
-    void deleteAll();
-}
