+++ src/main/java/com/hearthstone/hearthstonedeckstorage/HearthstoneDeckStorageApplication.java
@@ -1,0 +1,13 @@
+package com.hearthstone.hearthstonedeckstorage;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HearthstoneDeckStorageApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(HearthstoneDeckStorageApplication.class, args);
+    }
+
+}
