+++ src/main/java/com/hearthstone/hearthstonedeckstorage/api/HearthstoneClient.java
@@ -1,0 +1,11 @@
+package com.hearthstone.hearthstonedeckstorage.api;
+
+import com.hearthstone.hearthstonedeckstorage.dto.CardDto;
+
+import java.util.List;
+
+public interface HearthstoneClient {
+
+    List<CardDto> getCardsBySet(String set);
+
+}
