+++ src/main/java/com/hearthstone/hearthstonedeckstorage/api/token/OAuth2FlowHandler.java
@@ -1,0 +1,9 @@
+package com.hearthstone.hearthstonedeckstorage.api.token;
+
+import com.hearthstone.hearthstonedeckstorage.api.token.model.TokenResponse;
+
+public interface OAuth2FlowHandler {
+
+    TokenResponse getToken();
+    
+}
