+++ src/main/java/com/hearthstone/hearthstonedeckstorage/api/token/model/TokenResponse.java
@@ -1,0 +1,12 @@
+package com.hearthstone.hearthstonedeckstorage.api.token.model;
+
+import lombok.Data;
+
+import java.time.Instant;
+
+@Data
+public class TokenResponse {
+    private String tokenType;
+    private String accessToken;
+    private Instant expiresIn;
+}
