+++ src/main/java/com/hearthstone/hearthstonedeckstorage/common/config/EnvConfig.java
@@ -1,0 +1,10 @@
+package com.hearthstone.hearthstonedeckstorage.common.config;
+
+import lombok.Data;
+
+@Data
+public class EnvConfig
+{
+    public static final String TOKEN_URL_NAME = "https://oauth.battle.net/token";
+    public static final String BASE_URL_NAME = "https://eu.api.blizzard.com/hearthstone/cards";
+}
