+++ src/main/java/com/hearthstone/hearthstonedeckstorage/common/config/TemplateConfig.java
@@ -1,0 +1,21 @@
+package com.hearthstone.hearthstonedeckstorage.common.config;
+
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class TemplateConfig {
+
+    @Bean
+    public RestTemplateBuilder restTemplateBuilder() {
+        return new RestTemplateBuilder();
+    }
+
+    @Bean
+    public RestTemplate hearthstoneClientRestTemplate(RestTemplateBuilder builder) {
+        return builder.build();
+    }
+
+}
