+++ src/main/java/com/hearthstone/hearthstonedeckstorage/common/modelmapper/ModelMapperConfig.java
@@ -1,0 +1,26 @@
+package com.hearthstone.hearthstonedeckstorage.common.modelmapper;
+
+import lombok.RequiredArgsConstructor;
+import org.modelmapper.ModelMapper;
+import org.modelmapper.convention.NamingConventions;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import static org.modelmapper.config.Configuration.AccessLevel.PRIVATE;
+
+@Configuration
+@RequiredArgsConstructor
+public class ModelMapperConfig {
+
+    @Bean
+    public ModelMapper modelMapper(){
+        ModelMapper modelMapper = new ModelMapper();
+
+        modelMapper.getConfiguration()
+                .setFieldMatchingEnabled(true)
+                .setFieldAccessLevel(PRIVATE)
+                .setSkipNullEnabled(true)
+                .setSourceNamingConvention(NamingConventions.JAVABEANS_MUTATOR);
+        return modelMapper;
+    }
+}
