+++ src/main/java/com/hearthstone/hearthstonedeckstorage/dto/CardsRO.java
@@ -1,0 +1,17 @@
+package com.hearthstone.hearthstonedeckstorage.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class CardsRO {
+
+    private List<CardDto> cards;
+    private Integer cardCount;
+    private Integer pageCount;
+    private Integer page;
+
+}
