--- src/test/java/com/hearthstone/hearthstone_deck_storage/HearthstoneDeckStorageApplicationTests.java
@@ -1,13 +1,0 @@
-package com.hearthstone.hearthstone_deck_storage;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HearthstoneDeckStorageApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
