+++ src/test/java/com/hearthstone/hearthstonedeckstorage/HearthstoneDeckStorageApplicationTests.java
@@ -1,0 +1,13 @@
+package com.hearthstone.hearthstonedeckstorage;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class HearthstoneDeckStorageApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
